+++ src/test/java/cart/dao/JdbcProductDaoTest.java
@@ -1,6 +1,6 @@
 package cart.dao;
 
-import cart.entity.ProductEntity;
+import cart.domain.product.ProductEntity;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
